get_key_paths: Fix path searching with an easier algorithm
I didn't debug the old algorithm, but the new one finds more and shorter
paths that are valid, so the path pruning was too aggressive it seems.

For example for Andreas Larsson it finds

	from  Linus Torvalds <torvalds@kernel.org>
	   1  Junio C Hamano <gitster@pobox.com>
	   2  Geoff Levand <geoff@infradead.org>
	   3  Geert Uytterhoeven (Signing key) <geert@linux-m68k.org>
	   4  Andreas Larsson <andreas@gaisler.com>

in the kernel keyring that the old algorithm didn't find.

The new algorithm is also more deterministic as it emits the paths in
sort order (by certificate id). (Actually it's sorted by rowid of the
public keys, but both gnupg and sqlite don't shuffle during import in
practise.)

The new algorithm is also considerably simpler using
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm

Runtime is comparable to the algorithm before.

One side effect is also that the graph for the root key is corrected, it
was

	from  79BE3E4300411886  Linus Torvalds <torvalds@kernel.org>
	   1  79BE3E4300411886  Linus Torvalds <torvalds@kernel.org>

before.

Link: https://patch.msgid.link/20250519101429.1558947-11-u.kleine-koenig@baylibre.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>