export-keyring: Default to store "clean" public keys
There is no need to keep duplicated self signatures and signatures by keys
not in the keyring. The default behavior of gpg seems to have changed
regarding the first type of signatures as commit 671a0c921fe0 ("Periodic
update from keyservers") in the korg-pgpkeys repo introduced quite some of
these.

New key material is also only added cleaned, so it's consequent to also
let export-keyring clean the keys. A side effect is that signatures made by
deleted keys are dropped, too, but I consider that an upside.

An arguable downside of this change however is that not only duplicated
data is dropped, but also 3rd-party signatures over SHA-1 hashes. While
this is good in general, this will considerably damage the Web of Trust
in the kernel keyring, disconnecting 483 keys from Linus POV.

Link: https://patch.msgid.link/20250519101429.1558947-13-u.kleine-koenig@baylibre.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>