export-keyring: Skip graph generation if the trust path didn't change
It can happen that a key changes without affecting the trust paths. That
might be a new (or expired) subkey or a new signature that doesn't
matter for the trust paths. So always compare the trust txt file with
the new paths to prevent noop changes in the svg representation.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@baylibre.com>
Link: https://patch.msgid.link/204f68d615855ac880f12d2de046fc694466b497.1749831371.git.u.kleine-koenig@baylibre.com
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>